drfblog/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.3
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.0.0
ecdsa==0.18.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.6.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
